Refactor selfhost bundle wiring
